requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1

SQLAlchemy==2.0.36
alembic==1.14.0
pymysql==1.1.1
cryptography

pydantic==2.10.3
pydantic-settings==2.6.1

python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

python-multipart==0.0.12
jinja2==3.1.4

reportlab==4.2.5
pillow==10.4.0

requests>=2.31.0